Restructure the whole core to have viewModel
