+++ Notchly/App/AppEnvironment.swift
@@ -1,5 +1,5 @@
 //
-//  Notchly+Environment.swift
+//  AppEnvironment.swift
 //  Notchly
 //
 //  Created by Mason Blumling on 4/19/25.
